+++ app/services/google_calendar_integrator.rb
@@ -7,15 +7,15 @@ class GoogleCalendarIntegrator
     @client = @google_account.client
   end
 
-  def sync_tasks
-    if @google_integration.calendar_integration? && @client
-      tasks = @google_integration.account_list.tasks.future.uncompleted.of_type(@google_integration.calendar_integrations)
-      tasks.map { |task| sync_task(task.id) }
-    end
+  def sync_data
+    return unless @google_integration.calendar_integration? && @client
+    tasks = @google_integration.account_list.tasks.future.uncompleted.of_type(@google_integration.calendar_integrations)
+    tasks.map { |task| sync_task(task.id) }
   end
+  alias sync_tasks sync_data
 
   def sync_task(task_id)
-    return nil unless @google_integration.calendar_integration? && @google_integration.calendar_id
+    return unless @google_integration.calendar_integration? && @google_integration.calendar_id
 
     task = Task.find_by(id: task_id)
     google_event = GoogleEvent.find_by(google_integration_id: @google_integration.id,